+++ src/common/ray.h
@@ -22,47 +22,47 @@ class ray {
             : orig(origin), dir(direction), tm(0)
         {}
 
-        __host__ __device__ ray(const point3& origin, const vec3& direction, double time)
+        __host__ __device__ ray(const point3& origin, const vec3& direction, float time)
             : orig(origin), dir(direction), tm(time)
         {}
 
         __host__ __device__ point3 origin() const  { return orig; }
         __host__ __device__ vec3 direction() const { return dir; }
-        __host__ __device__ double time() const    { return tm; }
+        __host__ __device__ float time() const    { return tm; }
 
-        __host__ __device__ point3 at(double t) const {
+        __host__ __device__ point3 at(float t) const {
             return orig + t*dir;
         }
 
     public:
         point3 orig;
         vec3 dir;
-        double tm;
+        float tm;
 };
 
 class ray_cuda {
     public:
         ray_cuda() {}
-        __host__ __device__ ray_cuda(const double3& origin, const double3& direction)
+        __host__ __device__ ray_cuda(const float3& origin, const float3& direction)
             : orig(origin), dir(direction), tm(0)
         {}
 
-        __host__ __device__ ray_cuda(const double3& origin, const double3& direction, double time)
+        __host__ __device__ ray_cuda(const float3& origin, const float3& direction, float time)
             : orig(origin), dir(direction), tm(time)
         {}
 
-        __host__ __device__ double3 origin() const  { return orig; }
-        __host__ __device__ double3 direction() const { return dir; }
-        __host__ __device__ double time() const    { return tm; }
+        __host__ __device__ float3 origin() const  { return orig; }
+        __host__ __device__ float3 direction() const { return dir; }
+        __host__ __device__ float time() const    { return tm; }
 
-        __host__ __device__ double3 at(double t) const {
+        __host__ __device__ float3 at(float t) const {
             return orig + t*dir;
         }
 
     public:
-        double3 orig;
-        double3 dir;
-        double tm;
+        float3 orig;
+        float3 dir;
+        float tm;
 };
 
 #endif
